Attempts manual Integer parsing
We could try this with floats too.
But the idea is you could catch more errors becuase you'd know exactly
which character caused the error. Would need to read about IEEetc

Also the current implementation allows for multiple `"."` in theory
which we could fix. We currently just leave it up to the float parsing
for now.