Refactors!
* Breaks out nested modules to be their own files.
* optimizes the binary patterns to re-use the match context.

The second point was interesting! We initially broke out the parse_until
function to be in its own module but doing so means the match context
can't be re-used, so we left it where it was.

We also removed the calls to inspect when raising an error for the same
reason, instead opting to return tuples that get matched on higher up
and raise the appropriate message there instead.