Implements some improvements discovered in benchmarking
We found adding a guard for an integer and putting that case first
seemed to be slightly slower than leaving things as they were.

We also found that calling String.trim was removing our ability to
re-use the match context and that we didn't have to call it if we added
a case to the start of parse_args which detected a space and chopped it off. We could trivially add more whitespace cases to that too but from benchmarking this worked out faster than calling String.trim a lot