Simplifies the fake source code generation
We remove the old approach of starting at the top of the tree and
instead create a random S expression that we then "wrap" in another one.

For now this means it generates source code that always looks like this:

(fn 0_or_more_non_s_expression_args s_expression 0_or_more_non_s_expression_args)
It would be better if each argument position could be any of the valid
args, but it gets tricky to ensure that doesn't become an infinite loop.